Split register fields into components
