+++ components/register/englishNameField.tsx
@@ -1,0 +1,39 @@
+/* eslint-disable react/jsx-props-no-spreading */
+import React from "react";
+import { Form } from "react-bulma-components";
+
+const { Input, Field, Control, Label } = Form;
+
+interface Props {
+  englishName: string;
+  setEnglishName?: (value: string) => void;
+  isAdmin: boolean;
+}
+
+const EnglishNameField: React.FunctionComponent<Props> = ({
+  englishName,
+  setEnglishName = () => {},
+  isAdmin = false,
+}: Props) => {
+  return (
+    <Field>
+      <Label>English Name</Label>
+      <Control>
+        <Input
+          value={englishName}
+          onChange={(event: React.ChangeEvent<HTMLInputElement>): void =>
+            setEnglishName(event.target.value)
+          }
+          disabled={!isAdmin}
+          required
+        />
+      </Control>
+    </Field>
+  );
+};
+
+EnglishNameField.defaultProps = {
+  setEnglishName: () => {},
+};
+
+export default EnglishNameField;
